backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# Azure Storage
azure-storage-blob==12.19.0
django-storages[azure]==1.14.2

# Razorpay
razorpay==1.4.1

# Email
django-anymail[sendgrid]==10.2

# Image Processing
Pillow==10.2.0

# Environment Variables
python-dotenv==1.0.0
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0

# Filtering
django-filter==23.5

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Security
django-ratelimit==4.1.0

# CORS
whitenoise==6.6.0

# Development
black==23.12.1
flake8==7.0.0
